moved DispatcherCommand implementations to separate file
